refactor: Move Nat building code to Helpers
